Atualiza o tipo do campo 'coverage_radius' no modelo StoreLocation de 'float' para 'integer', garantindo a consistência dos dados e melhorando a definição do modelo.

+++ app/Models/StoreLocation.php
@@ -34,7 +34,7 @@ final class StoreLocation extends BaseModel
         'store_id' => 'integer',
         'latitude' => 'float',
         'longitude' => 'float',
-        'coverage_radius' => 'float',
+        'coverage_radius' => 'integer',
         'is_main' => 'boolean',
         'is_active' => 'boolean',
         'created_at' => 'datetime',
